requirements.txt
gradio
transformers
Pillow
torchvision
scikit-image